Moving WndProc to Game class
As Application class is already big enough to also contain all the EVENTs functions, designed the Game class to stack all Event funcs and EventLoop handling.

But then I need to move WndProc() func to the Game class and register WndProc() there too, as if not I'd need to declare all the virtual funcs for Events in the Application - that I don't want to do.

So moving everything to the Game class.